// Downloads/HavenBridge-master/frontend/src/types/models.ts
export interface RecentActivity {
  type: string;
  date: string;
  description: string;
  socialWorkerName: string;
}

export interface ImpactOverview {
  totalResidents: number;
  totalSessions: number;
  totalDonations: number;
  activeSafehouses: number;
}